feat: external JSON RPC model support
This commit adds initial support non python-based models which
are spawned as separate threads which communicate via stdin/stdout
using a simple JSON RPC API.

The RPC API is as follows:

  Sends the following JSON RPC as a single line of text to the external
  processes stdin:
    {"function" := "model_output" OR "model_state_update" OR "model_info" OR "model_update",
     "model" := parameter dictionary mapping string names to JSON-compatible primtives,
     "time" := JSON compatible number,
     "state" := list (or object with `tolist()` method) of JSON-compatible primitives,
     "input" := list JSON-compatible primitives}

  The external process should return a simple JSON compatible value.

  N.B., by "JSON compatible" we mean it trivially is convertable to/from
  a JSON string via json.dumps and json.loads.

The 'rpc_inv_pendulum' example illustrates this functionality
by simply making the previously native component models
from the 'inv_pendulum' example instead be separate python
processes which communicate via JSON over stdin/stdout.

The top-level TOML config file for a JSON RPC model should, for each
component which is operating as an external JSON RPC process,
include a "process" field which describes the executable/script to
run for that model (i.e., it should be a process that listens on
stdin for a single-line JSON RPC as described above, and return
a reply as a single-line JSON string on stdout) and may optionally
contain a "run_command" field specifying if the executable/script
should actually be an argument to some other executable (e.g.,
the 'rpc_inv_pendulum' example has each components' "run_command"
field set as "python3" and then specifies via the "process" field
the name of the python script to run for that JSON RPC component
model.